refactor(types): move append and extend into QueryLangSet
