+++ tests/unit/peapods/test_container.py
@@ -10,12 +10,7 @@ from jina.helper import random_name
 from jina.parser import set_pea_parser, set_ping_parser
 from jina.peapods.container import ContainerPea
 from jina.peapods.pea import BasePea
-from jina.proto import jina_pb2
-from jina.types.document import uid
-from jina.types.ndarray.generic import NdArray
-
 from tests import random_docs
-
 
 cur_dir = Path(__file__).parent
