+++ tests/unit/peapods/test_remote.py
@@ -1,7 +1,7 @@
 import pytest
 
 from jina.parser import set_gateway_parser, set_pea_parser
-from jina.peapods.pod import GatewayPod
+from jina.peapods.pod.gateway_pod import GatewayPod
 
 if False:
     from jina.peapods.remote import PeaSpawnHelper
